Use typed UUIDs for silo user and group
This commit changes SiloUser and SiloGroup to use typed UUIDs.

The biggest reason that this couldn't happen without a bunch of work was
the `lookup_resource` macro: for a resource like SshKey, it looked like

```
lookup_resource! {
    name = "SshKey",
    ancestors = [ "Silo", "SiloUser" ],
    lookup_by_name = true,
    soft_deletes = true,
    primary_key_columns = [ { column_name = "id", rust_type = Uuid } ]
}
```

One of the methods that the `lookup_resource` macro generates will
create a lookup for ancestors of the resource, and this was one using
the type returned from the corresponding authz resource's `id` method:

```
            quote! { .filter(dsl::#parent_id.eq(#parent_authz_name.id())) },
```

Changing SiloUser to use a typed UUID in the authz resource:

```
authz_resource! {
    name = "SiloUser",
    parent = "Silo",
    primary_key = { uuid_kind = SiloUserKind }, <-- here
    roles_allowed = false,
    polar_snippet = Custom,
}
```

and changing the `SiloUser` db model to use `DbTypedUuid` meant that
a call to `to_db_typed_uuid` was required. The lookup_resource macro has
no type information from the string "SiloUser", so this PR adds a check:
if the ancestor string is suffixed with a '*', then the lookup_resource
macro should assume that the `parent_id` is a typed UUID, and generate
the call to `to_db_typed_uuid`.

Most of the work after that was mechanical, changing Uuid to their typed
equivalent, changing method argument types, etc etc.

Some other related things made it into this PR:

- UserBuiltIn now also uses a typed UUID as well, distinguishing them
  from silo users

- Actor no longer has the `actor_id` method, instead requiring call
  sites to check which variant of Actor is being used

- AuthenticatedActor stores the full Actor instead of only the actor id,
  leading to typed comparisons in its oso::PolarClass impl

- User and Group path params are now typed