Audit log MVP (#7339)
Initial implementation of [RFD
523](https://rfd.shared.oxide.computer/rfd/0523).

## High-level design

* Logging an operation has two steps, corresponding to two app layer
methods called directly in the request handler:
* `audit_log_entry_init`: called before anything else, and if it fails,
we bail -- this guarantees nothing can happen without getting logged
* `audit_log_entry_complete`: called after the operation succeeds or
fails, filling in the row with the success or failure result. Currently
we only log the HTTP status code and possibly error message, but we will
fill this in further with, e.g., the ID of the created resource (if
applicable), and maybe the entire success response.
* This log is stored in CockroachDB and not somewhere else (like
Clickhouse) because we need an immediate guarantee at write time that
the audit log initialization happened before we proceed with the API
operation.
* The audit log can only be retrieved by fleet viewers at
`/v1/system/audit-log`
* The audit log list is powered by a SQL view that filters for only
completed entries
* The audit log list is ordered by `time_completed`, not `time_started`.
This turns out to be very important — see the doc comment on
`audit_log_list` in `nexus/db-queries/src/db/datastore/audit_log.rs`.
* Audit log entries have unique IDs in order to let clients deduplicate
them if they fetch overlapping ranges
* Timestamps could not be used as the primary key because (a) timestamp
collisions are possible, and (b) we are ordering by `time_completed`,
but not all entries in the audit log table have non-null
`time_completed`

## Operations logged

See `nexus/src/external_api/http_entrypoints.rs`. My goal was to start
by logging the operations that create sessions and tokens. Eventually I
think we want to log pretty much everything that's not a GET.

* `login_saml`: last step of SAML login, creates web session
* `login_local`: username/password login, creates web session
* `device_auth_confirm`: last step of token create
* `project_create` and `project_delete`
* `instance_create` and `instance_delete`
* `disk_create` and `disk_delete`

## Next steps

Things that are not in this PR, but which we will want to do soon,
possibly as soon as this release. I put the highest priority items
first.

### Log ID of created resource

For actions that create a resource, like disk or instance create, we
need to at least log the ID of the resource created. Even for token and
session creation, we can probably log the ID of the created token or
session. We may also want to log names if we have them.

### Log display name of user and silo

We only have UUIDs for user and silo and they are not very pleasant to
work with. It's a lot easier to see what's going on at a glance if we
have display names. On top of that, after a user or silo is deleted,
there isn't a way to look them up in the API by ID and get that info.

### Auto-complete uncompleted entries

Unlike with initialization (because we bail if it fails), we do not have
a guarantee that audit log completion runs successfully because we don't
want to turn every loggable operation into a saga to enable rollbacks.
To deal with this, we will likely need a background job to complete any
rows hanging around uncompleted for longer than N minutes or hours.
Because these will not have success or error info about the logged
operation, we will probably need an explicit third kind of completed
entry, like `success`/`error`/`timeout`.

### Versioned log format

We may want to indicate breaking changes to the log format so that
customers update whatever system is consuming and storing the log.

### Silo-level audit log endpoint

In this PR, the audit log can only be retrieved by fleet viewers at a
system-level endpoint. We will probably want to allow silo admins to
retrieve an audit log scoped to their silo. That will require

* A silo-scoped `/v1/audit-log` endpoint accessible only to silo admins
that does more or less what the system-level one does, plus `where
silo_id = <silo_id>`
* A `SiloAuditLog` authz resource alongside `AuditLog` that is tied to a
specific silo
* More robust logging of the silo an operation takes place in, probably
related to the above point about better actor logging on login actions.
The external authenticator actor is not in a silo, so currently we are
not writing down what silo a login attempt is happening in.

### Log putative user for login operations

For failed login attempts we want to know who they were _trying_ to log
in as. For SAML login this may not be meaningful as we only get the
request from the IdP after login was successful over there, but for
password login we could log the username.

### Log full JSON response

We may want to go as far as to log the entire JSON response. One minor
difficulty I ran into is that Dropshot handles serializing the response
struct to JSON, so we don't have access to the serialized thing in the
request handlers. Feels like a shame to serialize it twice, but we might
have to if we want to write down the response.

### Clean up old entries

Background task to delete entries older than N days, as determined by
our as-yet-undetermined our retention policy. We need to keep an eye on
how fast the table will grow, but it seems we already have some tables
that are quite huge compared to this one and we don't clean them up yet,
so I'm not too worried about it. We expect customers will want to
frequently fetch the log and save it off-rack, so the retention period
probably doesn't need to be very long.

### Log a bunch more events

Right now the audit log calls are a bit verbose. Dropshot deliberately
does not support middleware, which would let us do this kind of thing
automatically outside of the handlers. Finding a more ergonomic and less
noisy way of doing the audit logging and latency logging might require a
declarative macro.