[omdb] Add missing newline to `omdb db ereport show` (#8805)
As seen in today's demo hour, the use of `serde_json::to_writer_pretty`
doesn't include a trailing newline, so the closing `}` of the ereport
JSON is followed immediately by the shell prompt. This both looks bad
and makes copying/pasting the JSON annoyinger. This commit adds the
newline. Whoopsie.