[nexus] Add address lot view method. (#8752)
Part of
https://github.com/oxidecomputer/terraform-provider-oxide/issues/468.
The terraform resource expects to be able to create/read/delete address
lots, and so far nexus only supports create/list/delete. This patch adds
a new endpoint for viewing an address lot by name or id.

This is my first attempt to work on this repo, so let me know if this is
off-base @internet-diglett.