[nexus] allow abbreviating ereport 'class' as 'k' (#8804)
This was one of @cbiffle's ideas to save a few bytes of CBO. Since we
expect to encode a "class" key for every ereport, shortening the fields
that are included in every message has a meaningful impact on how much
we can buffer. Also, loss records (which are generated when the SP
restarts or when ereports don't fit in the buffer) lack a class string,
because the presence of a top-level "lost" key can be used to identify
them. So, we should handle that here, as well.

Also includes some SP sim improvements for testing this.