make sleds report their CPU families to Nexus (#8725)
RFD 505 proposes that instances should be able to set a "minimum
hardware platform" or "minimum CPU platform" that allows users to
constrain an instance to run on sleds that have a specific set of CPU
features available. This allows a user to opt a VM into advanced
hardware features (e.g. AVX-512 support) by constraining it to run only
on sleds that support those features. For this to work, Nexus needs to
understand what CPUs are present in which sleds.

Have sled-agent query CPUID to get CPU vendor and family information and
report this to Nexus as part of the sled hardware manifest.

----

This is largely code Greg had put together that, now that the control
plane bits that build on it are imminently to be PR'd, it's time to merge.

Co-authored-by: Greg Colombo <greg@oxidecomputer.com>