Planning reports (#8631)
Fixes #8284, fixes #8548.

Replaces many (but not yet all) planner log messages with entries in a
structured planning report. Those reports are comprised of "step
reports", each of which corresponds to a subroutine of `do_plan`. The
complete report may be formatted (in English) via a `Display`
implementation. If planning from the `blueprint_planner` background task
is successful, a report is attached to the
`BlueprintPlannerStatus::Targeted` variant so that it may be displayed
via `omdb`.