[MGS] Reject SP reset requests to our own sled (#8830)
Fixes #8768 by implementing "option 1" from that issue's proposed
solutions:

* Extends MGS's config file slightly to specify which _sled_ it's
running on (once it determines which scrimlet position it's in).
* Adds a config option to disallow SP reset requests to the local
scrimlet.
* Adds a custom dropshot error type for `sp_component_reset()` that can
cleanly report this particular kind of failure.

This is definitely the first custom dropshot error in MGS (although we
need some others: #8013, #8014, #8350), and maybe the first in omicron
more generally? The server side is pretty straightforward, I think, but
the client side ended up being a little messy. I'm not sure how much of
that is because of the rework I had to do to `try_all_serially()` vs the
inherent changes (e.g., the fact that we have a different progenitor
error type for this one endpoint).

I'll definitely put this on a racklette before merging, but this should
be close enough to review, and I'm very interested in feedback about
integrating the custom error type.