Store the silo admin group name
The silo admin group is a special group that is automatically created
during silo creation where members are granted the silo admin role. This
name however is not stored, but this is currently ok: any existing silo
won't have the ability to delete groups, meaning that the automatically
created admin group is there to stay and a group with a duplicate name
cannot be created.

Very soon there will be silos with provision types that _can_ delete
groups, meaning this name has to be known in order for Nexus to create
the appropriate policy when a group with a matching name is created
again. It was a mistake not to store this parameter with the silo, so
rectify that with this PR.