Add a user provision type to silo user and group
Silo users and groups have traditionally defined "external id" as
whatever the silo's identity provider (or in the case of ApiOnly silos,
an operator) assigns for them: this could be username, email, a random
id, etc.

SCIM throws one of many wrenches at us because the specification defines
an external id attribute for users and groups that is distinct from
other related attributes that could be sent, including username and
email. Importantly the SCIM client that is creating a user or group
could send a NULL external ID, which is not supported by the current
field's database type.

The first step to supporting SCIM's extra attributes is to add the
silo's user provision field to our user and group models, and use this
field to change the semantics of the external id field depending on what
provision type is used. For the existing ApiOnly and JIT silos this will
stay the same.

The second step that will help with type safety is to use this new user
provision type field to separate the silo user and group database models
into distinct silo user and group types at the datastore level, and have
separate enum variants for each type that wrap an inner typed object:

```rust
pub enum SiloUser {
    /// A User created via the API
    ApiOnly(SiloUserApiOnly),

    /// A User created during an authenticated SAML login
    Jit(SiloUserJit),
}
```

Call sites now have to use either the ApiOnly or Jit variant of this new
higher level silo user type. With the addition of user provision type to
the silo user model, Nexus can also now check that the provision type on
the user matches the silo, or (like with some of the local idp
functions) accept only one specific variant.

The third step that will also help with type safety is to have typed
lookups for silo users and groups, as different user provision types
will have different fields that they enforce uniqueness for: for
example, only search for ApiOnly and JIT users using external_id, but
for SCIM use username.

One possible bug that this commit addresses is that the creation of the
recovery silo did not have a check to ensure that the created silo was a
"LocalOnly" identity type. The code prior to this commit could create a
SamlJit silo, which is not ideal for a recovery silo!

Another possible bug potential addressed is a mismatch in the silo's
user provision type, and the users within that silo. This can be
enforced via functions that only accept one variant of the new higher
level enum, and is checked at the lower layers.

Another bug addressed: in `silo_group_ensure` there was a small window
between `silo_group_ensure_query` and unwrapping the result of
`silo_group_optional_lookup` where, if it were possible to delete
groups, a racing delete would cause the unwrap to panic.

Also! Fill in `test_ensure_same_silo_group`, why that was left truncated
is beyond me, but don't look at the git blame please.