Add boolean `fleet_viewer` and `silo_admin` to `/v1/me` response (#8861)
Dramatically simpler alternative to #8515. Rather than doing the
rigmarole required to get the actual specific role on the fleet and
silo, we just do little authz checks to get the specific things we care
about. The web console is the only consumer that cares about this
endpoint so it feels fine to do it that way.