[Reconfigurator] Fix planner skipping RoT / bootloader updates (add `TufArtifactMeta::board`) (#8872)
When trying Reconfigurator-based updates on a racklette today with a
real TUF repo, the planner skipped all the RoT bootloader and RoT
updates because none of the artifact `name`s matched the `board`s
reported in inventory. Our tests were assuming these were the same (and
in prod they _are_ the same on the SP), but in real TUF repos these
don't match: all the RoT and bootloaders report a `board` of
`oxide-rot-1`, but the artifact names are pretty varied.

This PR adds a new `board` value to the `tuf_artifact` table
and the `TufArtifactMeta` struct. Outside of tests, the changes are
almost identical to those from #8729 that added a `sign` field in this
same spot. This allows us to stop assuming that `artifact.name` ==
`inventory_caboose.board`, and instead check `artifact.board` ==
`inventory_caboose.board`. It depends on
https://github.com/oxidecomputer/tufaceous/pull/40, which changes the
way our fake data is generated to be more consistent with the state of
production devices and artifacts. This perturbs some of the
reconfiguartor-cli tests a bit, which significantly inflates the diff;
the actual code changes here are pretty small.