Relax networking requirements for instance start
Previously in `dpd_ensure` we were bailing out if we encounter
any errors while notifying dpd of nat changes. This means that
in the event of one of our switch zones being unavailable
we would no longer allow users to start instances, which is
probably a bit too strict.

Also, since the dendrite daemons have a NAT reconciliation RPW, we can
rely on it to catch us up instead of hard failing whenever there is
a failure to notify them of updates.