Take `TypedUuid`s out of the external API (#8910)
I noticed in
https://github.com/oxidecomputer/omicron/pull/8803#discussion_r2301588397
that there are some typed UUIDs leaking into the external API. As I said
in the comment, I don't think this information is likely to be useful
for clients because we don't actually use it in any of the request body
types, only the responses.

I don't think this change hurts correctness in our code because the
conversion to generic UUID usually happens right at the end of request
handling, when models are converted to views for response serialization.
However, there are a couple of spots where the struct is used a little
earlier in the process (see how the affinity datastore functions return
view structs) or is shared with internal API code, so there may be a
tiny risk to correctness.