Update kstat sampler futures when a target is updated (#8915)
- When a target is added to the `KstatSampler`, we also add a future
that resolves when we need to generate a sample from that target. Prior
to this, when a target was updated or removed, that future was never
touched! That means we could be polling for samples for removed targets
or that we might have multiple futures for the same target. This ensures
that we always update or remove the futures, along with the targets
themselves.
- Fixes #8889