TQ: cluster proptest Action/Event cleanup (#8899)
This builds on #8874 

To make using using tqdb easier, we put all necessary information inside
events. This works because of deterministic replay using the same
infrastucture as the proptest itself. To ensure we are truly
deterministically replaying we must assert that any Event in our log has
actually been generated when we go to apply it in tqdb. In the common
case case this is equivalent to asserting that the `DeliveredEnvelope`
is actually the same as the one pulled off the `bootstrap_network` vec
during runs. In order to guarantee this a few changes needed to be made:

1. Determinism exists, except for in the crypto code because we don't
seed the various random number generators, and therefore different key
shares and rack secrets get generated in different runs. We could seed
them, but then this makes it possible that we accidentally seed them in
production code. More importantly for our current purposes, though, is
that it's tedious and unnecessary. Instead we just implement comparison
methods for messages that ignore things like key shares. This works fine
because if the shares are not self-consistent with the rack secret and
the parameters such as threshold that we don't change the crypto code
itself will fail immediately.

2. We had a few actions that generated multiple events. Unfortunately,
applying these events would result in mutating the test networks
multiple times, resulting in a difference between which message was
recorded and which was actually pulled out to be delivered. This was
fixed by changing each action to only do a single thing at a time, like
deliver one envelope or commit a configuration at one node. This also
allows for finer grain interleaving and matches better with our TLA+
spec/model checking. This was observable by looking at the event logs.

The change to an action typically generating a single event means,
however, that we need to generate more actions per run to get the
equivalent functionality. A single action won't result in N commits or N
envelopes delivered. Therefore, each test run now needs to have
significantly more actions generated. Unfortunately this can make test
runs even longer. In order to help alleviate this we made three other
changes:

1. We change the invariant checks to only look at nodes that could
possibly be mutated by an event application. We call these the "affected
nodes". This prevents looping over every node in each check.
2. We reduce the member universe, and hence the size of the state space.
3. We reduce the total number of test cases per run.