[reconfigurator-cli] Extend target-release test to step through zone updates (#8940)
This is a step toward addressing #8478. It doesn't quite get there:

* I stopped at the point where we're ready to update Nexus; the current
behavior is wrong, and it didn't seem worth adding that wrong behavior
to this test.
* Our simulated system doesn't set up all zone types (it's missing
boundary NTP, cockroach, oximeter, and multinode clickhouse).

I'm not sure the second bullet is _super_ important for this test; we
don't really treat different zone types all that differently, other than
that some are expunge -> add and others are upgraded in place (and we do
have zones of both of those types in our simulated system). But I'd like
to leave #8478 open until we can address the first one.

In terms of review: the non-expectorate changes are quite small. I'd
recommend skimming the expectorate changes and checking the points where
there are comments about what just happened or is about to happen more
carefully. (As I was doing this I discovered that our simulated system
started with 3 pantry zones but during the upgrade test it expunged them
without replacing them; that's what led to changing the example system's
`target_crucible_pantry_zone_count`. So I think it is worth at least
skimming this output to see if there's anything else funky that I
missed.)