Report IP Pool utilization as capacity and remaining (#8928)
- Remove IP version-specific utilization types. All pools are only of
one version, so we can use the same types for both.
- Report IP Pool utilization through the API as a floating-point
capacity and count of _remaining_ addresses, rather than count of
allocated. This avoids dealing with enormous bit-width numbers like a
u128. The cost is reduced precision when either the capacity or
remaining is > 2**53, but in that case, the caller almost certainly
doesn't care about that. As the remaining number of addresses is
smaller, they get perfect precision.
- Fixes #8888
- Fixes #5347