[sled-agent] Add endpoints and `omdb` subcommands to stop and start the switch zone (#8959)
This is to assist with debugging and development around #8480;
specifically, it's item 2 on [the initial
plan](https://github.com/oxidecomputer/omicron/issues/8480#issuecomment-3033007065):

> De-risk the existing sled agent mechanism for tearing down the switch
zone in this case.
> a. add an internal API for it
> b. add an omdb command to invoke that API
> c. test this a bunch by putting the Sidecar into A2 and/or using
ignition to power it off and then using the omdb command to trigger the
response that we want

This PR adds `a` and `b` so we can do `c` on a racklette.