another RFD 314/PR transcription error: extended APIC space
short history here: this bit is set in hardware, the extended space is
not supported by bhyve, but the bit was passed through. hiding this bit
"should" be fine, but out of caution we're not doing that in the first
statically-defined CPU platform. this bit was not set in the MILAN_CPUID
blob, though, even though the RFD was updated.