sled-agent: begin incorporating Cosmo support
Extend the existing `Baseboard` type to incorporate Cosmo, and introduce
a new `OxideSled` enum to represent different Oxide sled types so that
a) we can differentiate between different kinds of sled, and b)
parameterize omicron code with constant data specific to a sled type.

Also, identify places where we call the `is_gimlet` method, and instead
call that, `is_oxide_sled` (none of these are actually specific to
Gimlets in particular).