No need for 3000 row tests (#8978)
LARGE_NUMBER_OF_ROWS was set to 3000 in order to confirm that the
queries related to user data export do not trigger full table scans for
a large number of rows, but it's only necessary to have a non-zero
amount of rows. Tune this down as it was causing tests to take a long
time.