[reconfigurator] Fix race in chicken switch loader
The chicken switch loading task exposes a watch channel that downstream
consumers can use to read the current value. It initializes that channel
to the `::default()` value of chicken switches and then reads the most
recent value from the db when it's activated, but that introduces a race
where consumers can see an incorrect value: if there is a chicken switch
value in the db, they can see the `::default()` _before_ the first
activation.

This PR closes the race window by setting the initial watch channel
value to `NotYetLoaded`, and then replacing that during the first
activation.