[reconfigurator] Fix race in chicken switch loader (#8998)
The chicken switch loading task exposes a watch channel that downstream
consumers can use to read the current value. It initializes that channel
to the `::default()` value of chicken switches and then reads the most
recent value from the db when it's activated, but that introduces a race
where consumers can see an incorrect value: if there is a chicken switch
value in the db, they can see the `::default()` _before_ the first
activation, allowing them to ignore the actual value in the db in favor
of whatever the default set is.

This PR closes the race window by setting the initial watch channel
value to `NotYetLoaded`, and then replacing that during the first
activation.

This fell out of trying to fix up tests on #8980, but is a legit bug and
easily separable from that work. So here it is!