Fix instance start saga tests
Before now the instant start saga tests weren't actually
driving ensuring that NAT configurations were eventually landing
on the dendrite instances. It turns out that Nexus in the
TestContext could not actually communicate with dendrite because
the port values were hard coded and our resolution logic relied on
comparing the different addresses for MGS (and all mgs instances
listen on localhost in testing). This commit introduces the
following changes:

* Use an updated version of Dendrite that allows us to provide a
  MGS SocketAddr on startup so it can use a simulated MGS to learn
  switch slot information
* Adjust the TestContext Builder logic to provide MGS information to
  dendrite
* Have Nexus ask dendrite for its switch slot information instead
  of having Nexus attempt to resolve it via MGS
* Change the test configuration to create a NIC for the instance so
  we actually exercise the NAT creation logic in the `dpd_ensure`
  saga node
* Introduce a test that ensures we still land configurations on the
  remaning switch even when one of our switches are unavailable