Make wicketd + installinator cosmo aware
This is the minimum set of changes to be able to update the next
generation sled Cosmo. The most intrusive change is choosing a
separate host phase 1 artifact.

This change does not attempt to differentiate between a Gimlet
and a Cosmo at the inventory level or ignition level which will
need to happen later via API changes.