hide FFXSR and WBNOINVD
FFXSR is used by Windows Server 2022 if available, which causes it to
die very early in boot - the feature is not supported by illumos bhyve
and we return a GP if the EFER bit is set.

WBNOINVD is also masked by default bhyve CPUID masking, though this one
is probably fine to pass through. Either way, because it is typically
masked, guests using this are basically untested (though as some
datapoints, Windows/Linux/illumos seem fine on bhyve with this bit
passed through..)