instance_migration_unknown_sled_type was legit flaky
when run on its own (at least on my workstation), the test instance is
always placed on the first test sled with a known CPU type. then,
migration fails because there is no available sled with a compatible
CPU, thus 507.

when run with other tests (such as... in CI), the test instance may be
placed on the unknown-CPU test sled, in which case it is unmigratable
regardless of capacity by virtue of its VMM having an unknown CPU
profile.